Add feature to add login to google
