+++ src/components/App/App.js
@@ -1,0 +1,15 @@
+import Login from '../Login/Login';
+import Logout from '../Logout/Logout';
+
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <Login />
+      <Logout />
+    </div>
+  );
+}
+
+export default App;
